requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
async-timeout==5.0.1
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
dateparser==1.2.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
narwhals==1.19.1
networkx==3.4.2
numdifftools==0.9.41
numpy==2.2.1
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.0.0
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
pyarrow==18.1.0
pycryptodome==3.21.0
pydeck==0.9.1
Pygments==2.18.0
python-binance==1.0.26
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scikit-learn==1.6.0
scipy==1.14.1
semopy==2.3.11
six==1.17.0
smmap==5.0.1
statsmodels==0.14.4
streamlit==1.41.1
sympy==1.13.1
ta==0.11.0
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
websockets==14.1
yarl==1.18.3
